// src/models/GoalScored/GoalScored.ts
export interface GoalScored {
    ball_last_touch: {
        player: string;
        speed: number;
    };
    goalspeed: number;
    impact_location:{
        X: number;
        Y: number;
    };
    scorer: {
        id: string;
        name: string;
        teamnum: number;
    };
    assister?: {
        name: string;
    }
}